Minor fixes
- move interface to its own file
- logg company house api errors
- fix typo
